MammoCrop/requirements.txt
pytorch
numpy
torchvision
scikit-learn
matplotlib
flwr